+++ server/src/main/java/com/hwc/server/service/CarFavoriteService.java
@@ -2,12 +2,16 @@ package com.hwc.server.service;
 
 import com.hwc.server.model.Car;
 import com.hwc.server.model.CarFavorite;
+import com.hwc.server.model.CarStats;
 import com.hwc.server.repository.CarFavoriteRepository;
 import com.hwc.server.repository.CarRepository;
+import com.hwc.server.repository.CarStatsRepository;
+import io.netty.util.internal.ObjectUtil;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
+import org.springframework.util.ObjectUtils;
 
 import java.util.Optional;
 
@@ -17,6 +21,7 @@ public class CarFavoriteService {
 
     private final CarRepository carRepository;
     private final CarFavoriteRepository carFavoriteRepository;
+    private final CarStatsService carStatsService;
 
     public Optional<CarFavorite> favorite(String userId, String carId) {
 
@@ -26,9 +31,14 @@ public class CarFavoriteService {
 
             Car carData = car.get();
 
-            if (carData.getFavoriteds().add(userId)) {
+            CarStats carStats = carStatsService.addFavorite(carData.getId(), userId);
 
-                carRepository.save(carData);
+            if (!ObjectUtils.isEmpty(carStats)) {
+
+                if (ObjectUtils.isEmpty(carData.getStats())) {
+                    carData.setStats(carStats);
+                    carRepository.save(carData);
+                }
 
                 return Optional.of(carFavoriteRepository.save(CarFavorite.builder()
                         .car(carData)
@@ -53,16 +63,17 @@ public class CarFavoriteService {
 
                 Car carData = car.get();
 
-                if (carData.getFavoriteds().remove(userId)) {
-                    carRepository.save(carData);
-                }
+                CarStats carStats = carStatsService.removeFavorite(carData.getId(), userId);
 
-                carFavoriteRepository.deleteById(carFavorite.get().getId());
+                if (!ObjectUtils.isEmpty(carStats)) {
+                    carFavoriteRepository.deleteById(carFavorite.get().getId());
+                }
 
             }
         }
 
     }
+
 
     public Page<CarFavorite> findAllByUserId(String userId, int page, int size) {
         return carFavoriteRepository.findAllByUserId(userId, PageRequest.of(page, size));
